pyproject.toml
[tool.poetry]
name = "fastapi-redis"
version = "0.1.0"
description = ""
authors = ["Neo <eyywqkgb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
redis-om = "^0.0.27"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^0.21.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
